backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.6.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
panda==0.3.1
pipfile==0.0.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.1
pytest==8.2.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
toml==0.10.2
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
